requirements.txt
streamlit
Pillow
numpy
matplotlib